--- a/intgemm/aligned.h
+++ b/intgemm/aligned.h
@@ -39,7 +39,9 @@ template <class T> class AlignedVector {
     }
 
     AlignedVector(const AlignedVector&) = delete;
+    AlignedVector(AlignedVector&) = delete;
     AlignedVector& operator=(const AlignedVector&) = delete;
+    AlignedVector& operator=(AlignedVector&) = delete;
 
     ~AlignedVector() {
 #ifdef _MSC_VER