--- a/intgemm/callbacks/implementations.inl
+++ b/intgemm/callbacks/implementations.inl
@@ -156,7 +156,7 @@ public:
     mult_reg = unquant_mult;
 #endif
     auto result = kernels::unquantize(input, mult_reg);
-    kernels::writePartial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
+    kernels::write_partial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
   }
 
 private:
@@ -184,7 +184,17 @@ public:
     auto result = kernels::relu<float>(kernels::unquantize(input, mult_reg));
     kernels::write(result, config.output_addr, info.row_idx * info.cols + info.col_idx);
   }
-  INTGEMM_TARGET void RunPartial(vi /*input*/, const OutputBufferInfo& /*info*/, Index /*partial*/) {}
+  INTGEMM_TARGET void RunPartial(vi input, const OutputBufferInfo& info, Index partial) {
+    // Workaround gcc 5 internal compiler error that can't read register members in debug.
+    vf mult_reg;
+#if !defined(__OPTIMIZE__) && (__GNUC__ == 5) && !defined(__clang__) && !defined(__INTEL_COMPILER)
+    asm ("vmovdqa %1, %0" : "=x" (mult_reg) : "m" (unquant_mult));
+#else
+    mult_reg = unquant_mult;
+#endif
+    auto result = kernels::relu<float>(kernels::unquantize(input, mult_reg));
+    kernels::write_partial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
+  }
 
 private:
   vf unquant_mult;
@@ -204,7 +214,10 @@ public:
     kernels::write(result, config.output_addr, info.row_idx * info.cols + info.col_idx);
   }
 
-  INTGEMM_TARGET void RunPartial(vi /*input*/, const OutputBufferInfo& /*info*/, Index /*partial*/) {}
+  INTGEMM_TARGET void RunPartial(vi input, const OutputBufferInfo& info, Index partial) {
+    auto result = kernels::add_bias_partial(input, config.bias_addr, info.col_idx, partial);
+    kernels::write_partial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
+  }
 
 private:
   AddBiasAndWrite config;
@@ -231,7 +244,18 @@ public:
     result = kernels::add_bias(result, config.bias_addr, info.col_idx);
     kernels::write(result, config.output_addr, info.row_idx * info.cols + info.col_idx);
   }
-  INTGEMM_TARGET void RunPartial(vi /*input*/, const OutputBufferInfo& /*info*/, Index /*partial*/) {}
+  INTGEMM_TARGET void RunPartial(vi input, const OutputBufferInfo& info, Index partial) {
+    // Workaround gcc 5 internal compiler error that can't read register members in debug.
+    vf mult_reg;
+#if !defined(__OPTIMIZE__) && (__GNUC__ == 5) && !defined(__clang__) && !defined(__INTEL_COMPILER)
+    asm ("vmovdqa %1, %0" : "=x" (mult_reg) : "m" (unquant_mult));
+#else
+    mult_reg = unquant_mult;
+#endif
+    auto result = kernels::unquantize(input, mult_reg);
+    result = kernels::add_bias_partial(result, config.bias_addr, info.col_idx, partial);
+    kernels::write_partial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
+  }
 private:
   vf unquant_mult;
   UnquantizeAndAddBiasAndWrite config;
@@ -259,7 +283,19 @@ public:
     result = kernels::relu<float>(result);
     kernels::write(result, config.output_addr, info.row_idx * info.cols + info.col_idx);
   }
-  INTGEMM_TARGET void RunPartial(vi /*input*/, const OutputBufferInfo& /*info*/, Index /*partial*/) {}
+  INTGEMM_TARGET void RunPartial(vi input, const OutputBufferInfo& info, Index partial) {
+    // Workaround gcc 5 internal compiler error that can't read register members in debug.
+    vf mult_reg;
+#if !defined(__OPTIMIZE__) && (__GNUC__ == 5) && !defined(__clang__) && !defined(__INTEL_COMPILER)
+    asm ("vmovdqa %1, %0" : "=x" (mult_reg) : "m" (unquant_mult));
+#else
+    mult_reg = unquant_mult;
+#endif
+    auto result = kernels::unquantize(input, mult_reg);
+    result = kernels::add_bias_partial(result, config.bias_addr, info.col_idx, partial);
+    result = kernels::relu<float>(result);
+    kernels::write_partial(result, config.output_addr, info.row_idx * info.cols + info.col_idx, partial);
+  }
 private:
   vf unquant_mult;
   UnquantizeAndAddBiasAndWriteRelu config;