--- a/intgemm/kernels/implementations.inl
+++ b/intgemm/kernels/implementations.inl
@@ -24,26 +24,31 @@ namespace intgemm {
 namespace kernels {
 
 /*
- * Write
+ * Write. Potentially unaligned memory, hence use storeu
  */
 CPU_ATTR static inline void write(vi input, int8_t* output, Index offset) {
-  *reinterpret_cast<vi*>(output + offset) = input;
+  storeu_ps(reinterpret_cast<float *>(output + offset), *reinterpret_cast<vf *>(&input));
+  //*reinterpret_cast<vi*>(output + offset) = input;
 }
 
 CPU_ATTR static inline void write(vi input, int16_t* output, Index offset) {
-  *reinterpret_cast<vi*>(output + offset) = input;
+  storeu_ps(reinterpret_cast<float *>(output + offset), *reinterpret_cast<vf *>(&input));
+  //*reinterpret_cast<vi*>(output + offset) = input;
 }
 
 CPU_ATTR static inline void write(vi input, int* output, Index offset) {
-  *reinterpret_cast<vi*>(output + offset) = input;
+  storeu_ps(reinterpret_cast<float *>(output + offset), *reinterpret_cast<vf *>(&input));
+  //*reinterpret_cast<vi*>(output + offset) = input;
 }
 
 CPU_ATTR static inline void write(vf input, float* output, Index offset) {
-  *reinterpret_cast<vf*>(output + offset) = input;
+  storeu_ps(reinterpret_cast<float *>(output + offset), input);
+  //*reinterpret_cast<vf*>(output + offset) = input;
 }
 
 CPU_ATTR static inline void write(vd input, double* output, Index offset) {
-  *reinterpret_cast<vd*>(output + offset) = input;
+  storeu_ps(reinterpret_cast<float *>(output + offset), *reinterpret_cast<vf *>(&input));
+  //*reinterpret_cast<vd*>(output + offset) = input;
 }
 
 /*